+++ hashtagmap-batch/src/test/java/com/songpapeople/hashtagmap/instagram/processor/InstagramBatchProcessorTest.java
@@ -1,4 +1,4 @@
-package com.songpapeople.hashtagmap.job;
+package com.songpapeople.hashtagmap.instagram.processor;
 
 import com.songpapeople.hashtagmap.place.domain.model.Place;
 import com.songpapeople.hashtagmap.scheduler.CrawlingResult;
