+++ hashtagmap-batch/src/test/java/com/songpapeople/hashtagmap/instagram/writer/InstagramBatchWriterTest.java
@@ -1,4 +1,4 @@
-package com.songpapeople.hashtagmap.job;
+package com.songpapeople.hashtagmap.instagram.writer;
 
 import com.songpapeople.hashtagmap.dto.CrawlingDto;
 import com.songpapeople.hashtagmap.dto.PostDto;
